Fix wron ussage of enums for declarations
Hi again!

I found another small bug. With the enums definitions it was forcing me to do this:

```
import {EmbedType} from 'geteventstore-promise';
...
this.client.persistentSubscriptions.getEvents(subscriptionName, streamName, this.messagesToGet, EmbedType.Body)
```

This was giving me the error: Cannot find Body of undefined.

I updated the enums to be types and in that way enforce the user to send once of the allowed values.